RBA-61: Migrationg and seeding for PG (#98)
* Update sqlx to 0.6 and fix spelling mistakes

* Add database seeding functionality

* Remove postgres.sql file since everything was moved into migrations

* Run linter and other things for web-api

* Update volume configuration for postgres in docker

+++ services/web-api/src/types/product.rs
@@ -91,8 +91,8 @@ impl Product {
             LIMIT $1 OFFSET $2
         "#,
         )
-        .bind(limit)
-        .bind(offset)
+        .bind(limit as i32)
+        .bind(offset as i32)
         .bind(user_id)
         .bind(ProductTypeArray(types))
         .bind(format!("%{}%", filter))
@@ -113,8 +113,8 @@ impl Product {
             LIMIT $1 OFFSET $2
         "#,
         )
-        .bind(limit)
-        .bind(offset)
+        .bind(limit as i32)
+        .bind(offset as i32)
         .bind(user_id)
         .bind(ProductTypeArray(types))
         .bind(format!("%{}%", filter))
@@ -137,8 +137,8 @@ impl Product {
             LIMIT $1 OFFSET $2
         "#,
         )
-        .bind(limit)
-        .bind(offset)
+        .bind(limit as i32)
+        .bind(offset as i32)
         .bind(user_id)
         .bind(ProductTypeArray(types))
         .bind(format!("%{}%", filter))
